refactor: Article,EditableBlock,BlockController 컴포넌트 분리
